requirements.txt
altgraph==0.17.2
future==0.18.2
pefile==2021.9.3
pip==21.1.2
pygame==2.0.2
pyinstaller-hooks-contrib==2021.3
pyinstaller==4.5.1
PyOpenGL-accelerate==3.1.5
PyOpenGL==3.1.5
pywin32-ctypes==0.2.0
setuptools==57.0.0
wheel==0.36.2
